backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

# Data & Storage
pyairtable==2.3.3

# AI
google-generativeai==0.8.3

# Document Processing
pypdf==5.1.0
python-docx==1.1.2

# Security
python-dotenv==1.0.1
python-magic==0.4.27
slowapi==0.1.9

# HTTP Client
httpx==0.28.1
